# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API Title",
        "description": "Your API Description",
        "contact": {
            "email": "contact@example.com"
        },
        "license": {
            "name": "License Name",
            "url": "http://www.example.com/license"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/exercise/user-plan-exercise": {
            "get": {
                "summary": "Retrieve user plan exercises with pagination 15 items per page",
                "description": "tags={\"Exercise\"},",
                "operationId": "aabb6d6c99afc17150f2bdd119794319",
                "responses": {
                    "200": {
                        "description": "User plan exercises retrieved successfully"
                    }
                }
            },
            "post": {
                "tags": [
                    "Exercise"
                ],
                "summary": "Create a new user plan exercise",
                "operationId": "75c169ef691c7b322487d441e00c5984",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "plan_id": {
                                        "description": "Must exist in exercise plans table",
                                        "type": "integer"
                                    },
                                    "user_id": {
                                        "description": "Must exist in users table",
                                        "type": "integer"
                                    },
                                    "status": {
                                        "type": "boolean"
                                    },
                                    "is_work": {
                                        "type": "boolean"
                                    },
                                    "days": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Note exercise created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NoteExercise"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/exercise/user-plan-exercise/{id}": {
            "get": {
                "summary": "Retrieve a specific note exercise",
                "operationId": "924ce65d17eec812550c0040c1c5390d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Note exercise's ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Note exercise retrieved successfully"
                    }
                }
            },
            "put": {
                "summary": "Update a specific user plan exercise",
                "operationId": "c8d62d74c3c500eaec796a801cefb3b0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User plan exercise's ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "plan_id": {
                                        "description": "Must exist in exercise plans table",
                                        "type": "integer"
                                    },
                                    "user_id": {
                                        "description": "Must exist in users table",
                                        "type": "integer"
                                    },
                                    "status": {
                                        "type": "boolean"
                                    },
                                    "is_work": {
                                        "type": "boolean"
                                    },
                                    "days": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User plan exercise updated successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "delete": {
                "summary": "Delete a specific user plan exercise",
                "operationId": "3f23b7a2f9be64a3c8029300dd527ca0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User plan exercise's ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User plan exercise deleted successfully"
                    },
                    "404": {
                        "description": "User plan exercise not found"
                    }
                }
            }
        },
...
        "/api/exercise/user-plan-exercise/date": {
            "get": {
                "tags": [
                    "Exercise"
                ],
                "summary": "Retrieve a plan by a specific date",
                "operationId": "3bc1ef2ae61638eb7999d58eb2a5ea4e",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "description": "The date to retrieve the plan for",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2022-01-01"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plan by date retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Plan by date retrieved successfully"
                                        },
                                        "plan": {
                                            "$ref": "#/components/schemas/PlanExercise"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200-rest": {
                        "description": "Rest day",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Rest day today"
                                        },
                                        "plan": {
                                            "properties": {
                                                "rest_day": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation errors"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Server error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "NoteExercise": {
                "properties": {
                    "user_id": {
                        "description": "Must exist in users table",
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "content": {
                        "type": "string"
                    },
                    "status": {
                        "type": "boolean"
                    },
                    "exercise_plan_id": {
                        "description": "Must exist in exercise plans table",
                        "type": "integer"
                    },
                    "exercise_id": {
                        "description": "Must exist in exercises table",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "PlanExercise": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "plan_id": {
                        "type": "string",
                        "example": "bulk_plan"
                    },
                    "is_done": {
                        "type": "boolean",
                        "example": false
                    },
                    "status": {
                        "type": "integer",
                        "example": 1
                    },
                    "exercises_count": {
                        "type": "integer",
                        "example": 3
                    },
                    "exercises": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Exercise"
                        }
                    },
                    "note": {
                        "type": "string",
                        "nullable": true
                    },
                    "done": {
                        "type": "boolean",
                        "nullable": true
                    },
                    "rest_day": {
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            },
            "Exercise": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "shoulder"
                    },
                    "is_done": {
                        "type": "boolean",
                        "example": false
                    },
                    "status": {
                        "type": "integer",
                        "example": 1
                    },
                    "details": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ExerciseDetail"
                        }
                    },
                    "note": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ExerciseDetail": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "exercise_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "exercise": {
                        "type": "string",
                        "example": "shoulder"
                    },
                    "is_done": {
                        "type": "boolean",
                        "example": false
                    },
                    "sets": {
                        "type": "integer",
                        "example": 0
                    },
                    "rir": {
                        "type": "integer",
                        "example": 15
                    },
                    "reps": {
                        "type": "integer",
                        "example": 12
                    },
                    "rest": {
                        "type": "integer",
                        "example": 5
                    },
                    "weight": {
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "unit": {
                        "type": "string",
                        "nullable": true
                    },
                    "day_names": {
                        "type": "string",
                        "nullable": true
                    },
                    "is_full": {
                        "type": "boolean",
                        "example": true
                    },
                    "duration": {
                        "type": "integer",
                        "example": 10
                    }
                },
                "type": "object"
            }
        }
    }
}